Check Key Date value exists in _mapCulModsKeyDate
The function _mapCulModsKeyDate uses explode() to retrieve the
start date, and if it exists, the end date. Howeve, it turns out
that if explode is given an empty array to explode, it will
return a one-element array with 0 as the key and an empty
string as the value. So the original code was crashing because
it was using the result at index 0 as the start date.
Fix: added a check before the above code that just returns
null if there is nothing in the Key Date field.